+++ Src/Private/Get-VbrRequiredModule.ps1
@@ -5,7 +5,7 @@ function Get-VbrRequiredModule {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.0.2
+        Version:        0.6.1
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux