+++ Src/Private/Get-DiagBackupToHvProxy.ps1
@@ -5,7 +5,7 @@ function Get-DiagBackupToHvProxy {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.6.0
+        Version:        0.6.1
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -27,16 +27,9 @@ function Get-DiagBackupToHvProxy {
     process {
         try {
             $HyperVBackupProxy = Get-VbrBackupProxyInfo -Type 'hyperv'
-            if ($Dir -eq 'LR') {
-                $DiagramLabel = 'Hyper-V Backup Proxies'
-                $DiagramDummyLabel = ' '
-            } else {
-                $DiagramLabel = ' '
-                $DiagramDummyLabel = 'Hyper-V Backup Proxies'
-            }
             if ($HyperVBackupProxy) {
                 $ProxiesAttr = @{
-                    Label = $DiagramLabel
+                    Label = 'Hyper-V Backup Proxies'
                     fontsize = 18
                     penwidth = 1.5
                     labelloc = 't'
@@ -44,28 +37,19 @@ function Get-DiagBackupToHvProxy {
                     style = 'dashed,rounded'
                 }
                 SubGraph MainSubGraph -Attributes $ProxiesAttr -ScriptBlock {
-                    # Dummy Node used for subgraph centering
-                    Node DummyHyperVProxy @{Label = $DiagramDummyLabel; fontsize = 18; fontname = "Segoe Ui Black"; fontcolor = '#005f4b'; shape = 'plain' }
-                    if ($Dir -eq "TB") {
-                        Node HvLeft @{Label = 'HvLeft'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain'; fillColor = 'transparent' }
-                        Node HvLeftt @{Label = 'HvLeftt'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain'; fillColor = 'transparent' }
-                        Node HvRight @{Label = 'HvRight'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain'; fillColor = 'transparent' }
-                        Edge HvLeft, HvLeftt, DummyHyperVProxy, HvRight @{style = $EdgeDebug.style; color = $EdgeDebug.color }
-                        Rank HvLeft, HvLeftt, DummyHyperVProxy, HvRight
-                    }
                     foreach ($ProxyObj in $HyperVBackupProxy) {
                         $PROXYHASHTABLE = @{}
                         $ProxyObj.psobject.properties | ForEach-Object { $PROXYHASHTABLE[$_.Name] = $_.Value }
                         Node $ProxyObj -NodeScript { $_.Name } @{Label = $PROXYHASHTABLE.Label; fontname = "Segoe Ui" }
-                        Edge -From DummyHyperVProxy -To $ProxyObj.Name @{constraint = "true"; minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                        Edge -From MainSubGraph:s -To $ProxyObj.Name @{constraint = "true"; minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                     }
                     Rank $HyperVBackupProxy.Name
                 }
 
                 if ($Dir -eq 'LR') {
-                    Edge $BackupServerInfo.Name -To DummyHyperVProxy @{minlen = 3 }
+                    Edge $BackupServerInfo.Name -To 'MainSubGraph' @{minlen = 3 }
                 } else {
-                    Edge $BackupServerInfo.Name -To DummyHyperVProxy @{minlen = 3 }
+                    Edge $BackupServerInfo.Name -To 'MainSubGraph' @{minlen = 3 }
                 }
             }
         } catch {