+++ Src/Private/Get-DiagBackupToRepo.ps1
@@ -5,7 +5,7 @@ function Get-DiagBackupToRepo {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.6.0
+        Version:        0.6.1
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -28,32 +28,16 @@ function Get-DiagBackupToRepo {
         try {
             $BackupRepo = Get-VbrBackupRepoInfo
             $LocalBackupRepo = Get-VbrBackupRepoInfo | Where-Object { $_.Role -like '*Local' }
-            $RemoteBackupRepo = Get-VbrBackupRepoInfo | Where-Object { $_.Role -like 'Dedup*' }
+            $DedupBackupRepo = Get-VbrBackupRepoInfo | Where-Object { $_.Role -like 'Dedup*' }
             $ObjStorage = Get-VbrBackupObjectRepoInfo
             $ArchiveObjStorage = Get-VbrBackupArchObjRepoInfo
+            $NASBackupRepo = Get-VbrBackupRepoInfo | Where-Object { $_.Role -like '*Share' }
 
             if ($BackupServerInfo) {
-                if ($Dir -eq 'LR') {
-                    $DiagramLabel = 'Backup Repositories'
-                    $DiagramDummyLabel = ' '
-                } else {
-                    $DiagramLabel = ' '
-                    $DiagramDummyLabel = 'Backup Repository'
-                }
-
                 if ($BackupRepo) {
-                    SubGraph MainSubGraph -Attributes @{Label = $DiagramLabel; fontsize = 22; penwidth = 1; labelloc = 't'; style = 'dashed,rounded'; color = $SubGraphDebug.color } {
-                        # Node used for subgraph centering
-                        Node BackupRepository @{Label = $DiagramDummyLabel; fontsize = 22; fontname = "Segoe Ui Black"; fontcolor = '#005f4b'; shape = 'plain' }
-                        if ($Dir -eq "TB") {
-                            Node RepoLeft @{Label = 'RepoLeft'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain'; fillColor = 'transparent' }
-                            Node RepoLeftt @{Label = 'RepoLeftt'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain'; fillColor = 'transparent' }
-                            Node RepoRight @{Label = 'RepoRight'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain'; fillColor = 'transparent' }
-                            Edge RepoLeft, RepoLeftt, BackupRepository, RepoRight @{style = $EdgeDebug.style; color = $EdgeDebug.color }
-                            Rank RepoLeft, RepoLeftt, BackupRepository, RepoRight
-                        }
+                    SubGraph MainSubGraph -Attributes @{Label = 'Backup Repositories'; fontsize = 22; penwidth = 1; labelloc = 't'; style = 'dashed,rounded'; color = $SubGraphDebug.color } {
                         if ($LocalBackupRepo) {
-                            SubGraph LocalRepos -Attributes @{Label = 'Local Repository'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph LocalRepos -Attributes @{Label = 'Local Repositories'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 # Node used for subgraph centering
                                 Node LocalReposDummy @{Label = 'LocalReposDummy'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
                                 if (($LocalBackupRepo | Measure-Object).count -le 3) {
@@ -89,25 +73,64 @@ function Get-DiagBackupToRepo {
                                     }
                                 }
                             }
-                            Edge -From BackupRepository -To LocalReposDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                            Edge -From MainSubGraph:s -To LocalReposDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                         }
-                        if ($RemoteBackupRepo) {
-                            SubGraph RemoteRepos -Attributes @{Label = 'Deduplicating Storage Appliances'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
-                                Node RemoteReposDummy @{Label = 'RemoteReposDummy'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain' }
-                                if (($RemoteBackupRepo | Measure-Object).count -le 3) {
-                                    foreach ($REPOOBJ in ($RemoteBackupRepo | Sort-Object -Property Name)) {
+                        if ($NASBackupRepo) {
+                            SubGraph NasRepos -Attributes @{Label = 'NAS Repositories'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                                # Node used for subgraph centering
+                                Node NasReposDummy @{Label = 'NasReposDummy'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
+                                if (($NASBackupRepo | Measure-Object).count -le 3) {
+                                    foreach ($REPOOBJ in ($NASBackupRepo | Sort-Object -Property Name)) {
                                         $REPOHASHTABLE = @{}
                                         $REPOOBJ.psobject.properties | ForEach-Object { $REPOHASHTABLE[$_.Name] = $_.Value }
                                         Node $REPOOBJ -NodeScript { $_.Name } @{Label = $REPOHASHTABLE.Label; fontname = "Segoe Ui" }
                                     }
 
-                                    Edge -From RemoteReposDummy -To $RemoteBackupRepo.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                                    Edge -From NasReposDummy -To $NASBackupRepo.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                                 } else {
-                                    $Group = Split-array -inArray ($RemoteBackupRepo | Sort-Object -Property Name) -size 3
+                                    $Group = Split-array -inArray ($NASBackupRepo | Sort-Object -Property Name) -size 3
                                     $Number = 0
                                     while ($Number -ne $Group.Length) {
                                         $Random = Get-Random
-                                        SubGraph "RemoteBackupRepo$($Number)_$Random" -Attributes @{Label = ' '; style = $SubGraphDebug.style; color = $SubGraphDebug.color; fontsize = 18; penwidth = 1 } {
+                                        SubGraph "NasBackupGroup$($Number)_$Random" -Attributes @{Label = ' '; style = $SubGraphDebug.style; color = $SubGraphDebug.color; fontsize = 18; penwidth = 1 } {
+                                            $Group[$Number] | ForEach-Object {
+                                                $REPOHASHTABLE = @{}
+                                                $_.psobject.properties | ForEach-Object { $REPOHASHTABLE[$_.Name] = $_.Value }
+                                                Node $_.Name @{Label = $REPOHASHTABLE.Label; fontname = "Segoe Ui" }
+                                            }
+                                        }
+                                        $Number++
+                                    }
+
+                                    Edge -From NasReposDummy -To $Group[0].Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                                    $Start = 0
+                                    $NasRepoNum = 1
+                                    while ($NasRepoNum -ne $Group.Length) {
+                                        Edge -From $Group[$Start].Name -To $Group[$NasRepoNum].Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                                        $Start++
+                                        $NasRepoNum++
+                                    }
+                                }
+                            }
+                            Edge -From MainSubGraph:s -To NasReposDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                        }
+                        if ($DedupBackupRepo) {
+                            SubGraph RemoteRepos -Attributes @{Label = 'Deduplicating Storage Appliances'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                                Node RemoteReposDummy @{Label = 'RemoteReposDummy'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain' }
+                                if (($DedupBackupRepo | Measure-Object).count -le 3) {
+                                    foreach ($REPOOBJ in ($DedupBackupRepo | Sort-Object -Property Name)) {
+                                        $REPOHASHTABLE = @{}
+                                        $REPOOBJ.psobject.properties | ForEach-Object { $REPOHASHTABLE[$_.Name] = $_.Value }
+                                        Node $REPOOBJ -NodeScript { $_.Name } @{Label = $REPOHASHTABLE.Label; fontname = "Segoe Ui" }
+                                    }
+
+                                    Edge -From RemoteReposDummy -To $DedupBackupRepo.Name @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                                } else {
+                                    $Group = Split-array -inArray ($DedupBackupRepo | Sort-Object -Property Name) -size 3
+                                    $Number = 0
+                                    while ($Number -ne $Group.Length) {
+                                        $Random = Get-Random
+                                        SubGraph "DedupBackupRepo$($Number)_$Random" -Attributes @{Label = ' '; style = $SubGraphDebug.style; color = $SubGraphDebug.color; fontsize = 18; penwidth = 1 } {
                                             $Group[$Number] | ForEach-Object {
                                                 $REPOHASHTABLE = @{}
                                                 $_.psobject.properties | ForEach-Object { $REPOHASHTABLE[$_.Name] = $_.Value }
@@ -127,11 +150,11 @@ function Get-DiagBackupToRepo {
                                     }
                                 }
                             }
-                            Edge -From BackupRepository -To RemoteReposDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                            Edge -From MainSubGraph:s -To RemoteReposDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
 
                         }
                         if ($ObjStorage) {
-                            SubGraph ObjectStorage -Attributes @{Label = 'Object Repository'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph ObjectStorage -Attributes @{Label = 'Object Repositories'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 Node ObjectStorageDummy @{Label = 'ObjectStorageDummy'; style = $SubGraphDebug.style; color = $SubGraphDebug.color; shape = 'plain' }
                                 if (($ObjStorage | Measure-Object).count -le 3) {
                                     foreach ($STORAGEOBJ in ($ObjStorage | Sort-Object -Property Name)) {
@@ -166,10 +189,10 @@ function Get-DiagBackupToRepo {
                                     }
                                 }
                             }
-                            Edge -From BackupRepository -To ObjectStorageDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                            Edge -From MainSubGraph:s -To ObjectStorageDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
                         }
                         if ($ArchiveObjStorage) {
-                            SubGraph ArchiveObjectStorage -Attributes @{Label = 'Archive Object Repository'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
+                            SubGraph ArchiveObjectStorage -Attributes @{Label = 'Archive Object Repositories'; fontsize = 18; penwidth = 1.5; labelloc = 't'; style = 'dashed,rounded' } {
                                 Node ArchiveObjectStorageDummy @{Label = 'ArchiveObjectStorageDummy'; style = $EdgeDebug.style; color = $EdgeDebug.color; shape = 'plain' }
                                 if (($ArchiveObjStorage | Measure-Object).count -le 3) {
                                     foreach ($STORAGEArchiveOBJ in ($ArchiveObjStorage | Sort-Object -Property Name)) {
@@ -204,12 +227,12 @@ function Get-DiagBackupToRepo {
                                     }
                                 }
                             }
-                            Edge -From BackupRepository -To ArchiveObjectStorageDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
+                            Edge -From MainSubGraph:s -To ArchiveObjectStorageDummy @{minlen = 1; style = $EdgeDebug.style; color = $EdgeDebug.color }
 
                         }
                     }
 
-                    Edge -From $BackupServerInfo.Name -To BackupRepository @{minlen = 3 }
+                    Edge -From $BackupServerInfo.Name -To MainSubGraph @{minlen = 3 }
                 }
             }
         } catch {