+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.6.1] - Unreleased
 
+### Added
+
+- Add support for NAS Repository (Backup-to-Repository)
+
 ### Changed
 
 - Allow EDGE to connect between Subgraph Clusters
+
+### Fixed
+
+- Fix veeam module version detection
 
 ## [0.6.0] - 2024-05-09