+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2024-08-31
+
+### Added
+
+- Add support for NAS Repository (Backup-to-Repository)
+
+### Changed
+
+- Allow EDGE to connect between Subgraph Clusters
+- Update Diagrammer.Core minimum to v0.2.3
+
+### Fixed
+
+- Fix veeam module version detection
+
 ## [0.6.0] - 2024-05-09
 
 ### Changed